Bug 1738111 - Look up handler apps by file extension if one exists when enumerating handlers. r=Gijs,mtigley
We're getting false negatives when checking for handler apps (at least on
Windows) because we're not bringing file extensions in that process even when
we know at least one we could use, we're only looking up associations based on
MIME types. That isn't reliable on Windows, so this patch adds the first
extension that we have stored in the handlers database to the search
parameters, which makes us more likely to find a handler if one is registered.

Differential Revision: https://phabricator.services.mozilla.com/D130526

UltraBlame original commit: f7a9e5c1f8e2054dc63f7c2b01d657d6d612bb14