Bug 1740569 - NativeNtBlockSet_Write needs to receive AnnotationWriter.  r=gsvelto
Bug 1420363 changed the type of `WriterFn`'s parameter from `HANDLE` to
`AnnotationWriter&` and rewrote `InternalWriteNotes` in mozglue.dll, but
it didn't update `NativeNtBlockSet_Write` in firefox.exe, which is used
when the launcher process is enabled.

This patch applies a part of D46848 to `NativeNtBlockSet_Write`.

Differential Revision: https://phabricator.services.mozilla.com/D130921

UltraBlame original commit: 908a35db13484d97777d00ef54834846b9004705