Bug 1741114 - Remove tileview debugger r=gfx-reviewers,jrmuizel
Unfortunately this is not currently maintained (it's been broken
since the addition of sub_slice compositor surface support).

It's also complicating efforts underway to refactor the way that
picture cache dependencies are handled as part of the visibility
pass refactoring work.

For these reasons, let's remove this functionality. If we find
someone with time to maintain it in future, it's likely that we
could restore most of the viewer code, and update the picture
cache code in WR to output the appropriate information.

Differential Revision: https://phabricator.services.mozilla.com/D131105

UltraBlame original commit: 3a753938cf4363a1c84eaee73b3707ebc18c3330