Bug 1738325 - Uninstall expired themes from BuiltInThemes.jsm. r=dao
This patch allows us to uninstall expired themes. It includes the uninstall logic in BuiltInThemes.ensureBuiltInThemes(), which runs on idle after startup and when about:addons is opened. It also changes aboutaddons.js so we don't show a blank Colorways section in about:addons if there are none installed. Finally, I removed the hidden pref gate controlling the Colorways feature. I initially included that to account for the possibility that we wouldn't finish the feature in time for 94 and we would've needed to uplift a simple patch to disable the dfeature. Since the feature has now shipped, there is no longer a need for the pref.

After colorways expire, the user gets to keep the one they had installed going into the expiry, if any. This is information that would usually be associated with the profile. However, the themes are built-in (i.e. not in the profile). I get around this by saving a list of retained themes in a pref. This has drawbacks, such as making it possible to recover expired themes in about:config. I ran this by Product, and they're okay with it.

Differential Revision: https://phabricator.services.mozilla.com/D130517

UltraBlame original commit: 81b4750bca5d69c7fe45ba40952132a9ee1f1094