Bug 1632870 - [ci] Stop showing taskgraph generation logs in taskcluster test output, r=taskgraph-reviewers,jmaher
This was dumping the entire output of the generation *after* the pytest
failure. Which basically makes these tests impossible to work with (as you need
to scroll up forever to find the error).  If you had a previous test run in
your shell buffer prior, it was really difficult to find the start of the
current run.

Ideally we'd still have a way of seeing this ouput when requested.. maybe we
could save it to a file? But for now this is a *much* saner default. Plus if
additional context is needed, we can focus on re-writing the tests to provide
it in their error messages.

Differential Revision: https://phabricator.services.mozilla.com/D131280

UltraBlame original commit: 145c441094fb70ef3f037e5c9602fb1656f5c29f