Bug 1632870 - [taskgraph] Store a hash of the 'test-setting' object, r=jmaher
This will make it easy for consumers to check whether a set of tasks are on the
same test setting or not by comparing their hash.

This also uses a 'ReadOnlyDict' to ensure the setting isn't modified later.

Differential Revision: https://phabricator.services.mozilla.com/D131285

UltraBlame original commit: 5363cdd650bcf7a8f68e24521d44b5f866d47102