Bug 1737112 - Update instructions for adding git-cinnabar to PATH on Windows r=mhentges,firefox-build-system-reviewers,glandium g=firefox-build-system-reviewers
Updated the path value in the instructions from "<drive>:/" to "/<drive>/" for
Windows so that it works as intended when added to the MSYS profile. Also
added a copyable link to the expected msys/etc dir based on the MOZILLABUILD
environment variable to direct the user to where to apply the change.

(If the MOZILLABUILD variable isn't defined, supply the expected default instead)

Differential Revision: https://phabricator.services.mozilla.com/D129480

UltraBlame original commit: 362251834a83d1cd1ae8321836070c1aa2fe52ef