Bug 1741776 - Make sure <option> is a containing-block for positioned descendants. r=miko
We have other code to make sure that options themselves are in-flow, so
we really want abspos boxes not to escape the select element.

Other browsers don't allow having out-of-flow option children at all
(they seem to force descendants to be in-flow). However that seems
fairly more complicated to implement, and I don't think it particularly
matters how we deal with this as long as we deal with it in a
sorta-reasonable way.

Differential Revision: https://phabricator.services.mozilla.com/D131785

UltraBlame original commit: 1f2d586e5968cea4c31a646177bfc65a521bae22