Bug 1376091 - For MOZ_WIDGET_GTK, change shortcut for `cmd_selectAll` from Alt+A to Ctrl+A. r=masayuki
With this patch, we stop registering `Alt+A` as a shortcut key for "Select All" on Linux and register `Ctrl+A` instead, which is the default shortcut for the command on other GTK applications, Chromium, and Windows. `Alt+A` was also causing a bug for menubar navigation because it hijacks `Alt+A` access key. This patch does not stop Linux environments from registering `Alt+A` as a shortcut key for "Select All"; it just drops Gecko's additional, non-standard shortcut key definition, and defaults to `Ctrl+A` instead.

Differential Revision: https://phabricator.services.mozilla.com/D131062

UltraBlame original commit: 853874dee29b2a068a74b2a550151afaffd939ce