Bug 1742564: Undo changes to paths exposed on `SiteManager` interfaces r=glandium
In bug 1741893, the patch attempted to ensure that paths were all
case-sanitized at the boundary points of the `SiteManager` classes.
However, the mistake that was made is that some of the //public//
properties exposed by the classes are used by Mach commands, so changing
their case caused path incompatibility down-the-line.

To resolve the problem, this patch undoes all extra normcasing, and
scopes the fixes to 1741893 to the specific places where paths are
compared within `site.py`.

Differential Revision: https://phabricator.services.mozilla.com/D131926

UltraBlame original commit: 085539d439e70ee8d6197ff596bd23c7523b93dd