Bug 1721217 - Part 3: Fix tests that try to open URIs for download, r=nika,preferences-reviewers
Previously, the tests were waiting for the load event, which was being emitted
for the initial about:blank. With the pref enabled, this event is no longer
propagated/fired to the content process. Instead of notifying the content
process that it needs to emit a load event for about:blank, we can instead make
the tests wait for a STATE_STOP event.

Differential Revision: https://phabricator.services.mozilla.com/D126843

UltraBlame original commit: 3755cfbe22fe7a49a0ae121645030c703af6518e