Bug 1742779 - Clean up TlsContext.get in wasm. r=rhunt
Push code that recovers the context from thread-local storage up the
call stack so that the general rule is that a context is available.

Document this constraint by starting an SMDOC to enumerate coding
rules for the wasm subsystem.

In many cases, the context can be read from the Instance's tls rather
than via TlsContext.get(), so make this improvement.

Document this constraint as well: TlsContext.get() should only be
called from cold code that can't reasonably get the context in another
way.

After that, always pass the available context down the call stack from
Instance calls.  (This becomes more entertaining in WasmBuiltins.cpp,
where we would previously call CallingActivation to get the activation
and then get the context from the activation, but where
CallingActivation would itself get the cx from the TlsContext.
There's no need for this complexity; the caller of CallingActivation
should just pass in the context, and in situations where the activtion
was used only to get the context, the call can be removed completely.)

Annotate the few remaining calls to TlsContext.get() as to why they
are OK, namely, they are in cold code where it would be a hardship to
pass in a context from generated code, or they are in signal handlers
or simulator support code where there's no way to directly pass the
context.

Differential Revision: https://phabricator.services.mozilla.com/D132031

UltraBlame original commit: cab1b34e7d1857b74b1a25d4a11b2bfa7d262773