Bug 1741156 - Initialize FreeType library in GPU process on Android. r=aosmond
In order to render text using Skia (as webrender does for blob images)
we must ensure that the Freetype library has been initialized. In the
parent process this is done by gfxPlatform, but the GPU process does
not have a gfxPlatform so we should do so in GPUParent instead. We
already did this on Gtk, but this patch makes us do so on Android too.

Differential Revision: https://phabricator.services.mozilla.com/D131233

UltraBlame original commit: ab8777eb33731d4ddff34e520fa5ca69730c0401