Bug 1743347 - [devtools] Move whyPaused.* string to a ftl file in devtools/shared. r=bomsy,jdescottes,fluent-reviewers,flod
Since those strings can be consumed by the PausedDebuggerOverlay, on the server,
we need to put them in devtools/shared.
As this will create some work for the l10n team, we migrate them to Fluent so we're
feeding 2 birds with 1 scone.
The `<LocalizationProvider>` is added directly in `<WhyPaused>` (and not in `<App>`,
as it's messing up with the React context (See Bug 1743155).
Some snapshots are updated. It's unfortunate that we don't see the translated
strings there anymore, but we can't have nice things in Jest.

Differential Revision: https://phabricator.services.mozilla.com/D132260

UltraBlame original commit: 6eaac86a49224c5818897ddb47466da5c52f4a53