Bug 1743519 - Replace XDRIncrementalStencilEncoder. r=arai
In order to remove the dependency of JSScript.h on the XDR files, move what was
the XDRIncrementalStencilEncoder to JSScript.h. Since it only contains a single
pointer, remove the UniquePtr<T> around it to avoid a silly indirection.

This also removes a superfluous check for asmjs that is already in the caller.
Also remove the outdated linearize helper.

Depends on D132428

Differential Revision: https://phabricator.services.mozilla.com/D132429

UltraBlame original commit: 64cf7017f49f1ab6bffbd4f3f2273c160ce08ba3